Cleanup of old dashboard and search files
Axon supports dashboard and search import/export through the UI.    The new format is a json file that has all of the corresponding searches and dashboard layouts in them.  The Powershell scripts to import or export dashboards and searches are no longer needed.  Removing the old format to reduce confusion with the new feature in the UI.